% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afsc_roots.R
\name{strip_roots}
\alias{get_prefix}
\alias{get_suffix}
\alias{has_prefix}
\alias{has_suffix}
\alias{strip_roots}
\title{Air Force Specialty Code (AFSC) Prefixes/Suffixes}
\usage{
strip_roots(x, allow_X = FALSE)

get_prefix(x, allow_X = FALSE)

get_suffix(x, allow_X = FALSE)

has_prefix(x, allow_X = FALSE)

has_suffix(x, allow_X = FALSE)
}
\arguments{
\item{x}{A character vector containing AFSCs.}

\item{allow_X}{Logical indicating whether or not the "X" placeholder is allowed. Default is \code{FALSE}.}
}
\value{
\code{strip_roots} returns the original vector of AFSCs with all of
  the prefixes and suffixes (if any) removed. The functions \code{get_prefix} and
  \code{get_suffix} return, respectively, the prefix and suffix (if any) from
  each AFSC.
}
\description{
Extract or strip off AFSC prefixes and suffixes (if any).
}
\examples{
afscs <- c("9T000", "X1N371E", "123ABC")
strip_roots(afscs)
get_prefix(afscs)
get_suffix(afscs)
}

