% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/afsc_type.R
\name{afsc_type}
\alias{afsc_type}
\alias{is_enlisted}
\alias{is_officer}
\title{Officer/Enlisted Air Force Specialty Codes (AFSCs)}
\usage{
afsc_type(x, allow_X = FALSE)

is_enlisted(x, allow_X = FALSE)

is_officer(x, allow_X = FALSE)
}
\arguments{
\item{x}{A character vector containing AFSCs.}

\item{allow_X}{Logical indicating wether or not the \code{"X"} placeholder is
allowed. Default is \code{FALSE}.}
}
\value{
\code{afsc_type} returns the AFSC type (i.e., either \code{"officer"}
  or \code{"enlisted"}). The functions \code{is_enlisted} and
  \code{is_officer} are logical tests.
}
\description{
Determine which AFSCs are officer and which are enlisted.
}
\examples{
afscs <- c(enlisted = "9T000", officer = "92T0", bad = "A9T0X0K")
afsc_type(afscs)
is_enlisted(afscs, allow_X = TRUE)
is_officer(afscs)
}

