% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afsc-validate.R
\name{is_valid_afsc}
\alias{get_invalid_afscs}
\alias{get_valid_afscs}
\alias{is_valid_afsc}
\title{Valid/Invalid Air Force Specialty Codes (AFSCs)}
\usage{
is_valid_afsc(x, allow_X = FALSE)

get_invalid_afscs(x, allow_X = FALSE, unique_only = FALSE)

get_valid_afscs(x, allow_X = FALSE, unique_only = FALSE)
}
\arguments{
\item{x}{A character vector containing AFSCs.}

\item{allow_X}{Logical indicating wether or not the \code{"X"} placeholder is
allowed. Default is \code{FALSE}.}

\item{unique_only}{Logical indicating wether or not only the unique AFSCs
should be returned. Default is \code{FALSE}.}
}
\value{
Returns a logical vector of the same length as its argument \code{x},
  containing \code{TRUE} for those elements that are considered as valid
  AFSCs, and \code{FALSE} otherwise.
}
\description{
Determine which AFSCs are valid or invalid.
}

