% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afsc.R
\name{get_career_group}
\alias{get_career_group}
\title{Career Group}
\usage{
get_career_group(x, allow_X = FALSE)
}
\arguments{
\item{x}{A character vector containing AFSCs.}

\item{allow_X}{Logical indicating wether or not the \code{"X"} placeholder is
allowed. Default is \code{FALSE}.}
}
\description{
Extract the career group indicated by the AFSC.
}

