% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{is_alnum}
\alias{is_alnum}
\title{Alphanumeric}
\usage{
is_alnum(x)
}
\arguments{
\item{string}{An object that inherits from class \code{"character"}.}
}
\value{
Returns \code{TRUE} if all characters in the string are alphanumeric
  and there is at least one character, FALSE otherwise.
}
\description{
General test if characters in a string are alphanumeric.
}
\examples{
is_alnum(c("1", 3, "ab2", "1-s", " 3", "", " ", "a ", "12a", "1.2"))
}

