% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/afsc_patterns.R
\name{unique_patterns}
\alias{unique_patterns}
\title{Unique Patterns}
\usage{
unique_patterns(x, show_ws = TRUE, show_ws_as = "w", as_table = FALSE)
}
\arguments{
\item{x}{A character vector.}

\item{show_ws}{Logical indicating whether or not to display whitespace.}

\item{show_ws_as}{Character used to indicate whitespace. Defaults to "w".}
}
\description{
Extract unique patterns from a character vector.
}

