% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afsc_roots.R
\name{strip_roots}
\alias{get_prefix}
\alias{get_suffix}
\alias{strip_roots}
\title{Air Force Specialty Code (AFSC) Prefixes/Suffixes}
\usage{
strip_roots(x)

get_prefix(x)

get_suffix(x)
}
\arguments{
\item{x}{A character vector containing AFSCs.}
}
\value{
\code{strip_roots} returns the original vector of AFSCs with all of
  the prefixes and suffixes (if any) removed. The functions \code{get_prefix} and
  \code{get_suffix} return, respectively, the prefix and suffix (if any) from
  each AFSC.
}
\description{
Extract or strip off AFSC prefixes and suffixes (if any).
}
\examples{
afscs <- c("9T000", "X1N371E", "123ABC")
strip_roots(afscs)
get_prefix(afscs)
get_suffix(afscs)
}

