% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/afsc.R
\name{shred_afsc}
\alias{shred_afsc}
\title{Shredout}
\usage{
shred_afsc(x, level = 1L, allow_X = FALSE)
}
\arguments{
\item{x}{A character vector containing AFSCs.}

\item{level}{Integer specifying the depth the AFSCs should be shred to.
Default is \code{1L} meaning return the first character (which should be
career group for valid AFSCs without any prefix).}

\item{allow_X}{Logical indicating wether or not the \code{"X"} placeholder is
allowed. Default is \code{FALSE}.}
}
\description{
Shred different levels.
}
\details{
For example, \code{level = 1L} extracts the career group.
}

